Allow the report processor to run on PE
The PE puppetserver runs as `pe-puppet`, not `puppet`.

The $settings::user variable resolves to the user that the Puppet Master node is configured to run as. This patch assumes that you are either enforcing as the agent on the Puppet Master node, or that if you're enforcing on a compile master that it will be the same as the Master of Masters.